ml-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
openai>=1.54.0
sentence-transformers>=3.3.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
chromadb>=0.5.0
numpy>=1.24.0
pydantic>=2.10.0
httpx>=0.27.0
spacy>=3.7.0
pytest>=8.0.0
pytest-mock>=3.14.0
pytest-asyncio>=0.24.0
hypothesis>=6.92.1
